pyproject.toml
[project]
name = "nabla_vl"
version = "0.1.0"
description = "Buliding SOTA large Vision-Language Model (VLM)"
authors = [
    { name = "takuyashintate", email = "shintate@nablas.com" }
]
dependencies = [
    "pip>=24.2",
    "transformers>=4.48.2",
    "tokenizers>=0.20.0",
    "sentencepiece>=0.2.0",
    "shortuuid>=1.0.13",
    "accelerate>=1.0.0",
    "peft>=0.13.1",
    'bitsandbytes>=0.44.1 ; platform_system == "Linux"',
    "pydantic>=2.9.2",
    "markdown2[all]>=2.5.0",
    "numpy<2.0.0",
    "scikit-learn>=1.5.2",
    "gradio>=4.44.1",
    "gradio-client>=1.3.0",
    "requests>=2.32.3",
    "httpx>=0.27.2",
    "uvicorn>=0.31.0",
    "fastapi>=0.115.0",
    "einops>=0.8.0",
    "einops-exts>=0.0.4",
    "timm>=1.0.9",
    "datasets==2.18.0",
    "deepspeed>=0.15.1",
    "setuptools>=75.1.0",
    "torchvision>=0.19.1",
    "wandb<0.17",
    "trl>=0.11.2",
    "pynvml>=11.5.3",
    "loguru>=0.7.2",
    "sacrebleu>=2.4.3",
    "evaluate>=0.4.3",
    "sqlitedict>=2.1.0",
    "tenacity>=9.0.0",
    "hf-transfer>=0.1.8",
    "huggingface-hub[cli]>=0.26.2",
    "pytablewriter>=1.2.0",
    "jupyter>=1.1.1",
    "tqdm>=4.66.6",
    "black>=25.1.0",
    "isort>=6.0.0",
    "flake8>=7.1.1",
    "uv>=0.6.0",
    "pycocoevalcap>=1.2",
    "streamlit>=1.40.2",
    "schedulefree>=1.4",
    'aria2>=0.0.1b0 ; platform_system == "Linux"',
    "opencv-python>=4.10.0.84",
    "langdetect>=1.0.9",
    "levenshtein>=0.26.1",
    "click>=8.1.8",
    "openpyxl>=3.1.5",
    'decord>=0.6.0 ; platform_system == "Linux"',
    "pymupdf>=1.25.2",
    "kaggle>=1.6.17",
    "img2dataset>=1.45.0",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = ["torch"] 

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/nabla_vl"]

[tool.black]
line-length = 88

[tool.isort]
line_length = 88
